pyproject.toml
[project]
name = "sae-recursive-depth"
version = "0.0.0"
description = "Recursive meta-sparse-autoencoders at depths 1-3 on Gemma-2-2B and GPT-2 Small."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Aragorn Wang", email = "wangaragorn@gmail.com" }]

dependencies = [
    "pyyaml>=6.0",
    "torch>=2.3",
    "transformers>=4.45",
    "transformer-lens>=2.0",
    "sae-lens>=6.0",
    "huggingface-hub>=0.24",
    "wandb>=0.17",
    "numpy>=1.26",
    "scipy>=1.13",
    "pandas>=2.2",
    "matplotlib>=3.8",
    "tqdm>=4.66",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
# Keep the lockfile deterministic across Aragorn's workstation and any grader machine.
# See https://docs.astral.sh/uv/concepts/projects/
